chore(base): Move `RoomInfo` inside another module.
This patch moves the `RoomInfo` type and its implementations inside the
`room_info` module.
